refactor: reorganize modules and improve vscodeApi imports
* Update import paths to use @common alias for shared modules
* Remove unused VsCode API imports
* Extract document event listeners into separate function
* Move UI button handlers to uiHandlers module
* Split messageHandlers initialization into separate function
