+++ src/historyView/index.html
@@ -18,8 +18,8 @@
     <script type="importmap" nonce="${nonce}">
       {
         "imports": {
-          "./modules/vscodeApi.js": "${vscodeApiUri}",
-          "./modules/domHandlers.js": "${domHandlersUri}"
+          "./modules/domHandlers.js": "${domHandlersUri}",
+          "@common/vscodeApi.js": "${vscodeApiUri}"
         }
       }
     </script>
