+++ src/latex/formatter/texfmt.ts
@@ -26,7 +26,6 @@ export async function runTexFmt(filePath: string): Promise<boolean> {
     } else {
       command.push('--nowrap');
     }
-    command.push('--check');
     command.push(`"${filePath}"`);
 
     const result = await executeCommand(command, { channel: CHANNEL });
