+++ src/progressView/index.html
@@ -18,14 +18,14 @@
         "imports": {
           "split.js": "${splitJsUri}",
           "marked": "https://esm.sh/marked@9.1.5",
-          "./modules/vscodeApi.js": "${vscodeApiUri}",
           "./modules/stateManager.js": "${stateManagerUri}",
           "./modules/messageHandlers.js": "${messageHandlersUri}",
           "./modules/domHandlers.js": "${domHandlersUri}",
           "./modules/constants.js": "${constantsUri}",
           "./modules/logFormatters.js": "${logFormattersUri}",
-          "./webviewState.js": "${webviewStateUri}",
-          "./modules/templateUtils.js": "${templateUtilsUri}"
+          "@common/vscodeApi.js": "${vscodeApiUri}",
+          "@common/webviewState.js": "${webviewStateUri}",
+          "@common/templateUtils.js": "${templateUtilsUri}"
         }
       }
     </script>
