+++ src/webview/index.html
@@ -17,10 +17,10 @@
           "./modules/fileHandlers.js": "${fileHandlersUri}",
           "./modules/uiHandlers.js": "${uiHandlersUri}",
           "./modules/messageHandlers.js": "${messageHandlersUri}",
-          "./modules/vscodeApi.js": "${vscodeApiUri}",
           "./modules/utils.js": "${utilsUri}",
-          "./webviewState.js": "${webviewStateUri}",
-          "./modules/templateUtils.js": "${templateUtilsUri}",
+          "@common/vscodeApi.js": "${vscodeApiUri}",
+          "@common/webviewState.js": "${webviewStateUri}",
+          "@common/templateUtils.js": "${templateUtilsUri}",
           "sortablejs": "https://cdn.skypack.dev/sortablejs"
         }
       }
