+++ src/webview/modules/fileHandlers.js
@@ -1,6 +1,6 @@
-import { vscode } from './vscodeApi.js';
+import { vscode } from '@common/vscodeApi.js';
 import { saveState } from './stateManager.js';
-import { getWebviewState, updateWebviewState } from './webviewState.js';
+import { getWebviewState, updateWebviewState } from '@common/webviewState.js';
 import { MULTIPLE_SELECTIONS } from './constants.js';
 import {
   addEventListenerSafely,
