+++ src/webview/script.js
@@ -1,27 +1,19 @@
-import { vscode } from './modules/vscodeApi.js';
 import { restoreState, saveState } from './modules/stateManager.js';
-import { setupMessageHandlers } from './modules/messageHandlers.js';
+import {
+  setupMessageHandlers,
+  initializeDataRequests,
+} from './modules/messageHandlers.js';
 import {
   setupUIHandlers,
   updateAutoToggleState,
   updateToolConfigToggleState,
   autoResizeTextarea,
+  setupDocumentListeners,
 } from './modules/uiHandlers.js';
 
+// Initialize data requests when window loads
 window.onload = function () {
-  const dataRequests = [
-    'getTheme',
-    'requestInputFile',
-    'requestReferenceFile',
-    'requestAuxiliaryFile',
-    'requestMediaFile',
-    'requestRecentCommits',
-    'requestBaseFile',
-  ];
-
-  dataRequests.forEach((request) => {
-    vscode.postMessage({ command: request });
-  });
+  initializeDataRequests();
 
   // Set default state for new folders
   restoreState();
@@ -50,103 +42,17 @@ window.onload = function () {
   }
 };
 
+// Setup message handlers
 setupMessageHandlers();
 
+// Setup UI when DOM is loaded
 document.addEventListener('DOMContentLoaded', function () {
   setupUIHandlers();
-
-  // Pack and Clean button handlers
-  const packButton = document.getElementById('packButton');
-  const cleanButton = document.getElementById('cleanButton');
-  const compareButton = document.getElementById('compareButton');
-  const acceptButton = document.getElementById('acceptButton');
-
-  if (packButton) {
-    packButton.addEventListener('click', () => {
-      const agent = document.getElementById('agent')?.value;
-      if (agent) {
-        vscode.postMessage({ command: 'packFiles', agent });
-      }
-    });
-  }
-
-  if (cleanButton) {
-    cleanButton.addEventListener('click', () => {
-      const agent = document.getElementById('agent')?.value;
-      if (agent) {
-        vscode.postMessage({ command: 'cleanFiles', agent });
-      }
-    });
-  }
-
-  if (compareButton) {
-    compareButton.addEventListener('click', () => {
-      const baseFile = document.getElementById('baseFile')?.value;
-      const editedFile = document.getElementById('editedFile')?.value;
-
-      if (baseFile && editedFile) {
-        vscode.postMessage({
-          command: 'compare',
-          baseFile: baseFile,
-          editedFile: editedFile,
-        });
-      } else {
-        vscode.postMessage({
-          command: 'showInformationMessage',
-          text: 'Please select both base and edited files to compare',
-        });
-      }
-    });
-  }
-
-  if (acceptButton) {
-    acceptButton.addEventListener('click', () => {
-      const baseFile = document.getElementById('baseFile')?.value;
-      const editedFile = document.getElementById('editedFile')?.value;
-
-      if (baseFile && editedFile) {
-        vscode.postMessage({
-          command: 'acceptEdited',
-          baseFile: baseFile,
-          editedFile: editedFile,
-        });
-      } else {
-        vscode.postMessage({
-          command: 'showInformationMessage',
-          text: 'Please select both base and edited files to accept changes',
-        });
-      }
-    });
-  }
 
   // Update initial toggle states
   updateToolConfigToggleState();
   updateAutoToggleState();
 
-  // Close dropdowns when clicking outside
-  document.addEventListener('click', (e) => {
-    const toolConfigOptions = document.getElementById('toolConfigOptions');
-    const autoExtractOptions = document.getElementById('autoExtractOptions');
-    const toggleToolConfig = document.getElementById('toggleToolConfig');
-    const toggleAutoExtract = document.getElementById('toggleAutoExtract');
-
-    if (
-      !toggleToolConfig?.contains(e.target) &&
-      !toolConfigOptions?.contains(e.target)
-    ) {
-      if (toolConfigOptions) {
-        toolConfigOptions.style.display = 'none';
-        updateToolConfigToggleState();
-      }
-    }
-    if (
-      !toggleAutoExtract?.contains(e.target) &&
-      !autoExtractOptions?.contains(e.target)
-    ) {
-      if (autoExtractOptions) {
-        autoExtractOptions.style.display = 'none';
-        updateAutoToggleState();
-      }
-    }
-  });
+  // Setup document-level event listeners
+  setupDocumentListeners();
 });
